Separation from original repository.
This prototype has been developped through one semester as a demonstration of functional concepts
for game development and reasoning about world state. The history is lost for simplicity and not
leaking confidential information. It is still contained in the private repository for the whole
project. I can expose parts on interest.